% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map of accidents (from the Traffic Safety Administration's
Fatality Analysis Reporting System) for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state number as (or coercible into) an integer for which the map will be plotted}

\item{year}{A year value in \%Y format, the year for which the accidents will be plotted}
}
\value{
\code{NULL}. This function is used for it's side effect, namely a map of the accidents
    for a given state and year
}
\description{
This function plots a map of all accidents (except those with missing longitude and/or latitude information)
for a particular state and year.
}
\note{
The data for each year should be in format "accident_<year>.csv.bz2". If a state number is not recognized,
   an error will be generated. State number 2 (Alaska) cannot be plotted.
}
\examples{
\dontrun{fars_map_state(8, 2015)}

}
