% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file and convert it to tibble data format (tbl_df which inherits from data.frame)}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A path to a file. Files ending in .gz, .bz2, .xz, or .zip will be 
automatically uncompressed}
}
\value{
A tibble data frame. If the file cannot be found an error is generated and execution is halted
}
\description{
This function reads in a csv file, and if it exists converts it to a tibble data frame which 
inherits properties from data.frame class. The tibble data frame is more suitable in some cases 
than standard data frame, as it, among other things, doesn't convert strings to factors, keeps
orignal variable names (e.g., if they contain a space), uses a somewhat different print method, etc.
}
\note{
The file to be read needs to be a csv or a compressed csv. Other file formats cannot be read with this function.
}
\examples{
fars_read("data/accident_2013.csv")
fars_read("data/accident_2013.csv.bz2")
fars_read(list.files("data", full.names = TRUE)[1])

}
