% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct a filename for a dataset from the Traffic Safety Administration's Fatality Analysis Reporting System
for a particular year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year in \%Y format, e.g., \code{2013}. A vector of year values, e.g. \code{2013:2016}
can also be used}
}
\value{
A filename of the type "accident_<year>.csv.bz2", where year is the input. If a vector of year values
     is provided, the function will return a character vector of filenames
}
\description{
This function constructs a filename of the type "accident_<year>.csv.bz2" which can then be
passed to \code{\link{fars_read}} to be loaded in the workspace as a tibble.
}
\note{
Input should either be a 4-digit integer year number in \%Y format or
    coercible to one (e.g., a string \code{"2013"})
}
\examples{
make_filename(2013)
make_filename(2013:2016)

}
